requirements.txt
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.3
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cheroot==8.3.0
click==7.1.1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
Flask==1.1.1
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==3.0.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lmdb==0.98
lxml==4.5.0
mahotas==1.4.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
opencv-contrib-python==4.2.0.32
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
web.py==0.40
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
